Extract depth-related methods into their own trait.
